+++ CHANGELOG.md
@@ -1,5 +1,20 @@
   # node-red-contrib-zwave-js Change Log
 
+  - 7.1.0
+
+    **New Features**
+      - Added new **Advanced Transceiver Settings** dialog (and where supported by the ZWave radio in use)
+        - Region settings
+        - RF Power settings
+        - Backup/Restore NVM
+      - The Inclusion completed dialog, now displays the security class that was achieved.
+
+    **Changes**
+      - Increase UI timeout to 15s to account for slower responding systems
+      - Optimisations to firmware upload Mechanics.
+      - Bump ZWave JS to 9.3.0.
+
+
   - 7.0.3
 
     **Changes**